added food amount

+++ Calories.h
@@ -11,10 +11,10 @@ class Calories : public QObject
 public:
     Q_INVOKABLE Calories(QObject* parent = 0);
 
-    JENSON_PROPERTY_GETSET(int, TotalCalories)
-    JENSON_PROPERTY_GETSET(int, Carbs)
-    JENSON_PROPERTY_GETSET(int, Proteins)
-    JENSON_PROPERTY_GETSET(int, Fats)
+    JENSON_PROPERTY_GETSET(float, TotalCalories)
+    JENSON_PROPERTY_GETSET(float, Carbs)
+    JENSON_PROPERTY_GETSET(float, Proteins)
+    JENSON_PROPERTY_GETSET(float, Fats)
 };
 
 SERIALIZABLE(Calories, Calories)
